Fix onboarding (#86)
With the migration to `ethers`, we accidentally broke onboarding. We
had assumed that `window.ethereum` was always present.

The initialization steps that use `window.ethereum` have been moved
into the `initialization` method, and wrapped in a `try..catch` to
ensure te remainder of the intialization proceeds even if MetaMask is
not installed.