Fix `remote` option of deploy script (#91)
The `remote` option for the deploy script would not correctly set the
remote used for deployment. It is used for everything but the most
important step: the deploy step. For the `git push` that ultimately
deploys the new build, the remote origin was hard-coded as the MetaMask
test dapp repo.

A git command is now used to dynamically fetch the remote URL using the
remote name given instead.